+++ config/routes.php
@@ -6,9 +6,9 @@ use App\Action\Auth\UserLoginSubmitAction;
 use App\Action\Calendar\AddEventAction;
 use App\Action\Calendar\ListEventAction;
 use App\Action\Cities\ListCitiesAction;
-use App\Action\Consulti\GenerateConsultLinkAction;
-use App\Action\Consulti\ListConsultLinkAction;
-use App\Action\Consulti\PinCodeActionCheck;
+use App\Action\Consult\GenerateConsultLinkAction;
+use App\Action\Consult\ListConsultLinkAction;
+use App\Action\Consult\PinCodeActionCheck;
 use App\Action\Diary\DeleteDiaryAction;
 use App\Action\Diary\GetDiaryAction;
 use App\Action\Diary\UploadDiaryAction;
